+++ src/main/scala/com/atomist/rug/runtime/CommandHandler.scala
@@ -1,6 +1,7 @@
 package com.atomist.rug.runtime
 
 import com.atomist.param.ParameterValues
+import com.atomist.rug.runtime.js.interop.RugContext
 import com.atomist.rug.spi.Handlers.Plan
 
 /**
@@ -8,5 +9,5 @@ import com.atomist.rug.spi.Handlers.Plan
   */
 trait CommandHandler extends ParameterizedRug {
   val intent: Seq[String]
-  def handle(ctx: CommandContext, params: ParameterValues) : Option[Plan]
+  def handle(ctx: RugContext, params: ParameterValues) : Option[Plan]
 }