+++ src/main/scala/com/atomist/rug/runtime/js/BaseJavaScriptHandlerFinder.scala
@@ -2,19 +2,18 @@ package com.atomist.rug.runtime.js
 
 import com.atomist.param.{Parameter, Tag}
 import com.atomist.rug.runtime.Rug
-import com.atomist.rug.runtime.js.interop.JavaScriptHandlerContext
 import jdk.nashorn.api.scripting.ScriptObjectMirror
 
 /**
   * Some common things used to extract handlers from nashorn
   */
-abstract class BaseJavaScriptHandlerFinder[T <: Rug](ctx: JavaScriptHandlerContext)
+abstract class BaseJavaScriptHandlerFinder[T <: Rug]
   extends JavaScriptUtils
     with JavaScriptRugFinder[T] {
 
   override def find(jsc: JavaScriptContext): Seq[T] = {
     jsc.vars.collect {
-      case Var(_, handler) if isValidHandler(handler) && handler.getMember("__kind") == kind => extractHandler(jsc, handler, ctx)
+      case Var(_, handler) if isValidHandler(handler) && handler.getMember("__kind") == kind => extractHandler(jsc, handler)
     }.flatten
   }
 
@@ -32,7 +31,7 @@ abstract class BaseJavaScriptHandlerFinder[T <: Rug](ctx: JavaScriptHandlerConte
     obj.hasMember("__kind") && obj.hasMember("__name") && obj.hasMember("__description") && obj.hasMember("handle")
   }
 
-  protected def extractHandler(jsc: JavaScriptContext, handler: ScriptObjectMirror, ctx: JavaScriptHandlerContext): Option[T]
+  protected def extractHandler(jsc: JavaScriptContext, handler: ScriptObjectMirror): Option[T]
 
   protected def name(obj: ScriptObjectMirror): String = obj.getMember("__name").asInstanceOf[String]