+++ src/main/scala/com/atomist/rug/runtime/js/JavaScriptResponseHandler.scala
@@ -2,19 +2,15 @@ package com.atomist.rug.runtime.js
 
 import com.atomist.param.{Parameter, ParameterValues, ParameterizedSupport, Tag}
 import com.atomist.rug.InvalidHandlerResultException
-import com.atomist.rug.runtime.js.interop.JavaScriptHandlerContext
 import com.atomist.rug.runtime.{InstructionResponse, ParameterizedRug, ResponseHandler}
 import com.atomist.rug.spi.Handlers.Plan
 import jdk.nashorn.api.scripting.ScriptObjectMirror
 
 /**
   * Extract response handlers from a Nashorn instance
-  * @param ctx
   */
-class JavaScriptResponseHandlerFinder(
-                                       ctx: JavaScriptHandlerContext
-                                     )
-  extends BaseJavaScriptHandlerFinder[JavaScriptResponseHandler](ctx) {
+class JavaScriptResponseHandlerFinder
+  extends BaseJavaScriptHandlerFinder[JavaScriptResponseHandler] {
   /**
     *
     * The value of kind in JS
@@ -22,7 +18,7 @@ class JavaScriptResponseHandlerFinder(
     */
   override val kind: String = "response-handler"
 
-  override protected def extractHandler(jsc: JavaScriptContext, handler: ScriptObjectMirror, ctx: JavaScriptHandlerContext): Option[JavaScriptResponseHandler] = {
+  override protected def extractHandler(jsc: JavaScriptContext, handler: ScriptObjectMirror): Option[JavaScriptResponseHandler] = {
     Some(new JavaScriptResponseHandler(jsc, handler, name(handler), description(handler), parameters(handler), tags(handler)))
   }
 }