+++ src/main/scala/com/atomist/rug/runtime/plans/LocalInstructionRunner.scala
@@ -5,6 +5,7 @@ import com.atomist.project.edit._
 import com.atomist.project.generate.ProjectGenerator
 import com.atomist.project.review.ProjectReviewer
 import com.atomist.rug.runtime._
+import com.atomist.rug.runtime.js.interop.RugContext
 import com.atomist.rug.spi.Handlers.Instruction._
 import com.atomist.rug.spi.Handlers.Status.{Failure, Success}
 import com.atomist.rug.spi.Handlers.{Instruction, Response}
@@ -18,7 +19,7 @@ import com.atomist.rug.spi.RugFunctionRegistry
   */
 class LocalInstructionRunner(rugs: Seq[AddressableRug],
                              projectManagement: ProjectManagement,
-                             commandContext: CommandContext,
+                             rugContext: RugContext,
                              rugFunctionRegistry: RugFunctionRegistry = DefaultRugFunctionRegistry)
   extends InstructionRunner
   with PlanSupport{
@@ -59,7 +60,7 @@ class LocalInstructionRunner(rugs: Seq[AddressableRug],
               Response(Success, None, None, Some(reviewResult))
             })
           case Some(rug: CommandHandler) =>
-            val planOption = rug.handle(commandContext, parameters)
+            val planOption = rug.handle(rugContext, parameters)
             Response(Success, None, None, planOption)
           case Some(rug: ResponseHandler) =>
             callbackInput match {