+++ src/test/scala/com/atomist/rug/runtime/js/JavaScriptCommandHandlerTest.scala
@@ -3,8 +3,7 @@ package com.atomist.rug.runtime.js
 import com.atomist.param.{SimpleParameterValue, SimpleParameterValues}
 import com.atomist.project.archive.{AtomistConfig, DefaultAtomistConfig}
 import com.atomist.project.common.MissingParametersException
-import com.atomist.rug.runtime.CommandContext
-import com.atomist.rug.runtime.js.interop.{JavaScriptHandlerContext, jsPathExpressionEngine}
+import com.atomist.rug.runtime.js.interop.{LocalRugContext, jsPathExpressionEngine}
 import com.atomist.rug.ts.TypeScriptBuilder
 import com.atomist.source.{SimpleFileBasedArtifactSource, StringFileArtifact}
 import com.atomist.tree.TreeMaterializer
@@ -13,13 +12,12 @@ import org.scalatest.{FlatSpec, Matchers}
 class JavaScriptCommandHandlerTest extends FlatSpec with Matchers{
 
   val atomistConfig: AtomistConfig = DefaultAtomistConfig
-  val treeMaterializer: TreeMaterializer = TestTreeMaterializer
 
   val kitties = "ShowMeTheKitties"
   val kittyDesc = "Search Youtube for kitty videos and post results to slack"
   val simpleCommandHandler =  StringFileArtifact(atomistConfig.handlersRoot + "/Handler.ts",
     s"""
-       |import {HandleCommand, Instruction, Response, CommandContext, Plan, Message} from '@atomist/rug/operations/Handlers'
+       |import {HandleCommand, Instruction, Response, HandlerContext, Plan, Message} from '@atomist/rug/operations/Handlers'
        |import {CommandHandler, Parameter, Tags, Intent} from '@atomist/rug/operations/Decorators'
        |
        |@CommandHandler("ShowMeTheKitties","Search Youtube for kitty videos and post results to slack")
@@ -30,7 +28,7 @@ class JavaScriptCommandHandlerTest extends FlatSpec with Matchers{
        |  @Parameter({description: "his dudeness", pattern: "^.*$$"})
        |  name: string
        |
-       |  handle(ctx: CommandContext) : Plan {
+       |  handle(ctx: HandlerContext) : Plan {
        |    let pxe = ctx.pathExpressionEngine()
        |
        |    if(this.name != "el duderino") {
@@ -54,7 +52,7 @@ class JavaScriptCommandHandlerTest extends FlatSpec with Matchers{
 
   val simpleCommandHandlerWitPresentable =  StringFileArtifact(atomistConfig.handlersRoot + "/Handler.ts",
     s"""
-       |import {HandleCommand, MappedParameters, Message, Instruction, Response, CommandContext, Plan} from '@atomist/rug/operations/Handlers'
+       |import {HandleCommand, MappedParameters, Message, Instruction, Response, HandlerContext, Plan} from '@atomist/rug/operations/Handlers'
        |import {CommandHandler, Parameter, MappedParameter, Tags, Intent} from '@atomist/rug/operations/Decorators'
        |
        |@CommandHandler("ShowMeTheKitties","Search Youtube for kitty videos and post results to slack")
@@ -68,7 +66,7 @@ class JavaScriptCommandHandlerTest extends FlatSpec with Matchers{
        |  @MappedParameter(MappedParameters.REPO_OWNER)
        |  owner: string
        |
-       |  handle(ctx: CommandContext) : Plan {
+       |  handle(ctx: HandlerContext) : Plan {
        |    let pxe = ctx.pathExpressionEngine()
        |
        |    if(this.name != "el duderino") {
@@ -91,7 +89,7 @@ class JavaScriptCommandHandlerTest extends FlatSpec with Matchers{
 
   val simpleCommandHandlerExecuteInstruction =  StringFileArtifact(atomistConfig.handlersRoot + "/Handler.ts",
     """
-       |import {HandleCommand, HandleResponse, Message, Instruction, Response, CommandContext, Plan} from '@atomist/rug/operations/Handlers'
+       |import {HandleCommand, HandleResponse, Message, Instruction, Response, HandlerContext, Plan} from '@atomist/rug/operations/Handlers'
        |import {CommandHandler, ResponseHandler, Parameter, Tags, Intent} from '@atomist/rug/operations/Decorators'
        |
        |@CommandHandler("ShowMeTheKitties","Search Youtube for kitty videos and post results to slack")
@@ -102,7 +100,7 @@ class JavaScriptCommandHandlerTest extends FlatSpec with Matchers{
        |  @Parameter({description: "his dudeness", pattern: "^.*$"})
        |  name: string
        |
-       |  handle(ctx: CommandContext) : Plan {
+       |  handle(ctx: HandlerContext) : Plan {
        |
        |    let result = new Plan()
        |
@@ -133,13 +131,13 @@ class JavaScriptCommandHandlerTest extends FlatSpec with Matchers{
 
   it should "be able to schedule an Execution and handle its response" in {
     val rugArchive = TypeScriptBuilder.compileWithModel(SimpleFileBasedArtifactSource(simpleCommandHandlerExecuteInstruction))
-    val finder = new JavaScriptCommandHandlerFinder(new JavaScriptHandlerContext("XX", treeMaterializer))
+    val finder = new JavaScriptCommandHandlerFinder()
     val handlers = finder.find(new JavaScriptContext(rugArchive))
   }
 
   it should "extract and run a command handler" in {
     val rugArchive = TypeScriptBuilder.compileWithModel(SimpleFileBasedArtifactSource(simpleCommandHandler))
-    val finder = new JavaScriptCommandHandlerFinder(new JavaScriptHandlerContext("XX", treeMaterializer))
+    val finder = new JavaScriptCommandHandlerFinder()
     val handlers = finder.find(new JavaScriptContext(rugArchive))
 
     handlers.size should be(1)
@@ -148,12 +146,12 @@ class JavaScriptCommandHandlerTest extends FlatSpec with Matchers{
     handler.description should be (kittyDesc)
     handler.tags.size should be(3)
     handler.intent.size should be(2)
-    val plan = handler.handle(SimpleContext, SimpleParameterValues(SimpleParameterValue("name","el duderino")))
+    val plan = handler.handle(LocalRugContext(TestTreeMaterializer), SimpleParameterValues(SimpleParameterValue("name","el duderino")))
   }
 
   it should "parse messages containing Presentables" in {
     val rugArchive = TypeScriptBuilder.compileWithModel(SimpleFileBasedArtifactSource(simpleCommandHandlerWitPresentable))
-    val finder = new JavaScriptCommandHandlerFinder(new JavaScriptHandlerContext("XX", treeMaterializer))
+    val finder = new JavaScriptCommandHandlerFinder()
     val handlers = finder.find(new JavaScriptContext(rugArchive))
     handlers.size should be(1)
     val handler = handlers.head
@@ -161,25 +159,16 @@ class JavaScriptCommandHandlerTest extends FlatSpec with Matchers{
     handler.description should be (kittyDesc)
     handler.tags.size should be(3)
     handler.intent.size should be(2)
-    val plan = handler.handle(SimpleContext, SimpleParameterValues(SimpleParameterValue("owner","his dudeness"), SimpleParameterValue("name","el duderino")))
+    val plan = handler.handle(LocalRugContext(TestTreeMaterializer), SimpleParameterValues(SimpleParameterValue("owner","his dudeness"), SimpleParameterValue("name","el duderino")))
   }
 
   it should "throw exceptions if required parameters are not set" in {
     val rugArchive = TypeScriptBuilder.compileWithModel(SimpleFileBasedArtifactSource(simpleCommandHandler))
-    val finder = new JavaScriptCommandHandlerFinder(new JavaScriptHandlerContext("XX", treeMaterializer))
+    val finder = new JavaScriptCommandHandlerFinder()
     val handlers = finder.find(new JavaScriptContext(rugArchive))
     val handler = handlers.head
     assertThrows[MissingParametersException]{
-      handler.handle(SimpleContext, SimpleParameterValues.Empty)
+      handler.handle(LocalRugContext(TestTreeMaterializer), SimpleParameterValues.Empty)
     }
   }
-}
-
-object SimpleContext extends CommandContext{
-  /**
-    * Id of the team we're working on behalf of
-    */
-  override def teamId: String = "blah"
-
-  override def pathExpressionEngine: jsPathExpressionEngine = new jsPathExpressionEngine(this)
 }