+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### CHANGED
 
+-   **BREAKING** Only require context when running handlers
 -   **BREAKING** Removed CommandRegistry and related usages
 -   **BREAKING** Renamed old RugFunction to RugDslFunction
 -   New text node structure. See [docs/treenodes.md]().